refactor(game-engine): extract pass family handlers (S27.8.2)
Slice 2 de S27.8 (refactor monolith `actions.ts`). Apres l'extraction
des actions speciales en S27.8.1, ce slice s'attaque a la famille
pass : un groupe coherent de 5 handlers autour des actions de passe
et remise.

Extraction (`actions/pass-actions.ts`, 301l)
- `handlePass` : flow complet PASS avec branchements On the Ball,
  Animosity, Stunty pass-range, Running Pass.
- `handleOnTheBallMove` / `handleOnTheBallDecline` : reactions du
  coach defenseur avant le declenchement effectif de la passe (cycle
  vers `handlePass` reste interne au module).
- `handleHandoff` : flow HANDOFF avec Animosity et Running Pass S3.
- `handleThrowTeamMate` : flow THROW_TEAM_MATE.

Aucune logique modifiee : les 5 handlers sont copies a l'identique
puis re-exportes via `import { handle* } from './pass-actions'`.

Imports orphelins retires de `actions.ts`
- `hasAnimosityAgainst` / `checkAnimosity` (animosity)
- `executePass` / `executeHandoff` (passing — `getPassRange` +
  `canAttemptPassForRange` restent pour `getLegalMoves`)
- `canApplyRunningPass` / `canApplyRunningPassToHandoff` /
  `markRunningPassUsed` (running-pass)
- `executeThrowTeamMate` (throw-team-mate — `canThrowTeamMate` +
  `getThrowRange` restent pour `getLegalMoves`)
- `getOnTheBallReactivePlayers` / `executeOnTheBallMove` /
  `markOnTheBallUsed` (on-the-ball)

Resultat
- `actions.ts` 2595 -> 2393 lignes (-202 nettes, ~7.8% supplementaire ;
  cumul -313 vs origine 2706 = ~11.6% deja gagne).
- `pass-actions.ts` ajoute a la liste d'exclusion coverage (meme
  quirk source-map v8 que `actions.ts`).
- 4888 tests game-engine + 1755 tests server passent. Coverage
  25.33%, au-dessus du seuil 25%.

Reste pour atteindre la cible S27.8 (<= 600 lignes pour `actions.ts`) :
- handleBlock / choice handlers (BlockChoose / PushChoose /
  FollowUpChoose / RerollChoose / DumpOffChoose)
- handleBlitz
- handleFoul
- handleMove / handleNormalMove / handleDodge / handleDodgeRoll /
  handleBallPickup / handleLeap / handleEndTurn / handleEndPlayerTurn
